fix 6.cpp

+++ 5.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 using namespace std;
 int main(){
-    int x[3] = {-3, 0, 3};
+    double x[3] = {-3, 0, 3};
     double long y[3] = {0, 0, 0};
     int a = 4;
     for (int i=0; i < 3; ++i){
